requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.30
protobuf==3.19.4
pymongo==4.1.1
requests==2.27.1
psycopg2==2.9.3
selenium==3.141.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
typing_extensions==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
